new test_plopper.py test suite, move public plopper to benchmark folder, refactoring strings
